pyproject.toml
[tool.poetry]
name = "plannerfamylybudget"
version = "0.1.0"
description = ""
authors = ["Lizaveta Kaptsilava <126101406+ElaKaptsilava@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
sentry-sdk = { extras = ["django"], version = "^2.1.1" }
social-auth-app-django = "^5.4.1"
django-crispy-forms = "^2.1"
django-bootstrap4 = "^24.3"
pillow = "^10.3.0"
django-environ = "^0.11.2"
django-filter = "^24.2"
psycopg2-binary = "^2.9.9"
celery = "^5.4.0"
redis = "^5.0.6"
django-silk = "^5.1.0"
openai = "^1.35.9"
pyparsing = "^3.1.2"
pydot = "^2.0.0"
python-dateutil = "^2.9.0.post0"
stripe = "^10.5.0"
django-colorfield = "^0.11.0"


[tool.poetry.group.dev.dependencies]
djlint = "^1.34.1"
pre-commit = "^3.7.1"
black = "^24.4.2"
ruff = "^0.4.3"
bandit = "^1.7.8"
isort = "^5.13.2"
factory-boy = "^3.3.0"
conventional-pre-commit = "^3.2.0"
pre-commit-hooks = "^4.6.0"
safety = "^3.2.0"
coverage = "^7.5.1"
django-debug-toolbar = "^4.4.2"
django-extensions = "^3.2.3"
flower = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
